pyproject.toml
[tool.poetry]
name = "agents-behave"
version = "0.1.0"
description = ""
authors = ["Pedro Sousa <pedronsousa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
behave = "^1.2.6"
fastapi = "^0.110.0"
flake8 = "^7.0.0"
httpx = "^0.27.0"
langchain = "^0.1.9"
langchain-core = "^0.1.26"
langchain-openai = "^0.0.7"
langchain-together = "^0.0.2.post1"
pyhamcrest = "^2.1.0"
pytest = "^8.0.1"
python-dotenv = "^1.0.1"
ruff = "^0.2.2"
uvicorn = "^0.27.1"
openai = "^1.13.3"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
